Move PluginEntry and helpers to :build-logic:kotlin-dsl-shared-runtime
