+++ platforms/core-configuration/kotlin-dsl/src/main/kotlin/org/gradle/kotlin/dsl/support/IO.kt
@@ -27,20 +27,6 @@ inline fun <T : AutoCloseable, U> T.useToRun(action: T.() -> U): U =
     use { run(action) }
 
 
-/**
- * Appends value to the given Appendable and simple `\n` line separator after it.
- *
- * Always using the same line separator on all systems to allow for reproducible outputs.
- */
-internal
-fun Appendable.appendReproducibleNewLine(value: CharSequence = ""): Appendable {
-    assert('\r' !in value) {
-        "Unexpected line ending in string."
-    }
-    return append(value).append("\n")
-}
-
-
 internal
 fun File.isParentOf(child: File): Boolean =
     child.canonicalPath.startsWith(canonicalPath)
